+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.0] - 2025-09-25
+
+### Added
+
+- Support multi lang read out
+- Collapse and expand the summary result card
+
+### Changed
+
+- History deletion action using swipe gesture instead of long press
+- Improve article extraction accuracy
+
 ## [0.8.0] - 2025-09-21
 
 ### Added